requirements.txt
alembic
amqp
annotated-types
anyio
bcrypt
billiard
blinker
boto3
botocore
Brotli
cachelib
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
distlib
dnspython
email-validator
exceptiongroup
filelock
Flask
Flask-Bcrypt
Flask-Login
Flask-Mail
Flask-Migrate
Flask-Moment
Flask-OAuthlib
Flask-Session
Flask-SQLAlchemy
Flask-WTF
fonttools
google-auth
google-auth-oauthlib
greenlet
h11
html5lib
idna
importlib-metadata
itsdangerous
Jinja2
jmespath
kombu
Mako
MarkupSafe
mysql-connector-python
numpy
oauthlib
packaging
pandas
pathlib
Pillow
platformdirs
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic_core
pydyf
PyJWT
pyphen
python-dateutil
python-dotenv
python-multipart
pytz
requests
requests-oauthlib
rsa
s3transfer
six
slugify
sniffio
SQLAlchemy
SQLAlchemy-Utils
tinycss2
typing_extensions
tzdata
urllib3
uvicorn
vendor
vine
virtualenv
wcwidth
weasyprint
webencodings
Werkzeug
WTForms
zipp
zopfli
